refactor: reorganize the package structure
