+++ src/main/java/study/calculator/StringAddCalculator.java
@@ -1,4 +1,4 @@
-package study.step2;
+package study.calculator;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Objects;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import study.step2.exception.NumberConversionException;
+import study.calculator.exception.NumberConversionException;
 
 public class StringAddCalculator {
     private static final int ZERO = 0;
