+++ src/main/java/study/calculator/exception/NumberConversionException.java
@@ -1,4 +1,4 @@
-package study.step2.exception;
+package study.calculator.exception;
 
 public class NumberConversionException extends RuntimeException {
     public NumberConversionException(String message) {
