+++ src/test/java/study/calculator/StringAddCalculatorTest.java
@@ -1,4 +1,4 @@
-package study.step2;
+package study.calculator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
